+++ yast/request.py
@@ -7,6 +7,10 @@ import json
 import typing
 from urllib.parse import quote, unquote
 
+
+class ClientDisconnect(Exception):
+    pass
+
 class Request(Mapping):
     def __init__(self, scope: Scope, receive: Recevie = None):
         self._scope = scope
@@ -90,6 +94,8 @@ class Request(Mapping):
                 yield message.get('body', b'')
                 if not message.get('more_body', False):
                     break
+            elif message['type'] == 'http.disconnect':
+                raise ClientDisconnect()
 
     async def body(self):
         if not hasattr(self, "_body"):