+++ changelog.md
@@ -1,6 +1,7 @@
 # v0.1.2
 1. async file & response
 2. static files
+3. relative_url
 
 # v0.1.1
 1. descorators can aysnc func